+++ Liber/Controllers/Account Handling/LoginViewController.swift
@@ -51,6 +51,7 @@ class LoginViewController: UIViewController {
         
         hideHUD(view: self.view)
     }
+    
 }