+++ Liber/Controllers/Tabs/BookcaseViewController.swift
@@ -88,7 +88,7 @@ class BookcaseViewController: UIViewController, UITableViewDelegate, UITableView
             
             cell?.textLabel?.text = book.title
         } else {
-            //TODO: - Print no content found.
+            //TODO: - Print no content found. https://stackoverflow.com/questions/28532926/if-no-table-view-results-display-no-results-on-screen
         }
         
         return cell!